Remove obsolete MapController and OrbitController "proxy" classes (#1769)
