--- docs/api-reference/map-state.md
@@ -1,1 +1,0 @@
-# MapState Class (Experimental)
