--- docs/api-reference/orbit-state.md
@@ -1,1 +1,0 @@
-# OrbitState Class (Experimental)
